github2wordpress/requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Deprecated==1.2.13
idna==3.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
python-wordpress-xmlrpc==2.3
requests==2.27.1
soupsieve==2.3.2
urllib3==1.26.9
wrapt==1.14.0
